change to pages router, refactoring, fix parsing errors
