+++ packages/frontend/src/pages/_app.tsx
@@ -1,9 +1,14 @@
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
 import { appWithTranslation } from 'next-i18next'
+import Layout from "@/components/layout/layout";
 
 function App({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
+  return (
+    <Layout>
+      <Component {...pageProps} />
+    </Layout>
+  )
 }
 
 export default appWithTranslation(App)
