+++ packages/frontend/src/pages/calendar/index.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import TitleContainer from "@/UI/TitleContainer/TitleContainer";
+import TitleContainer from "@/shared/components/TitleContainer/TitleContainer";
 import React from "react";
 
 function index() {
