+++ packages/frontend/src/pages/table/index.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import TitleContainer from "@/UI/TitleContainer/TitleContainer";
+import TitleContainer from "@/shared/components/TitleContainer/TitleContainer";
 import React from "react";
-import TableContainer from "./_components/TableContainer/TableContainer";
+import TableContainer from "@/components/table/TableContainer/TableContainer";
 import { useTranslation } from "next-i18next";
-import ActionContainer from "./_components/ActionContainer/ActionContainer";
+import ActionContainer from "@/components/table/ActionContainer/ActionContainer";
 
 function Page() {
   const { t } = useTranslation('table');
