+++ packages/frontend/src/shared/components/StyledButton/StyledButton.tsx
@@ -1,4 +1,4 @@
-import React, { ButtonHTMLAttributes, PropsWithChildren } from "react";
+import React from "react";
 import { ButtonContainer } from "./StyledButton.styles";
 import { ButtonProps } from "antd/es/button/button";
