+++ packages/frontend/src/shared/components/TitleContainer/TitleContainer.style.ts
